refactor: LongTermComparator를 ScheduleComparator로 변경
